+++ src/Handlers/Events/ScriptableEventHandler.php
@@ -207,7 +207,7 @@ class ScriptableEventHandler
                 return $model;
             }
         } catch (\Exception $ex) {
-            \Log::error('Error occurred while loading event script. ', $ex->getMessage());
+            \Log::error('Error occurred while loading event script. ' . $ex->getMessage());
         }
 
         return null;