+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - DF-1222 Removing write to database from script caching so that lookups are not processed prematurely
 
+## [0.6.2] - 2017-10-30
+### Fixed
+- Fix typo in error log output handling
+
 ## [0.6.1] - 2017-09-18
 ### Added
 - DF-1177 & DF-1161 Added services for GitHub and GitLab with linking to server side scripting
@@ -56,7 +60,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-11-30
 First official release working with the new [dreamfactory](https://github.com/dreamfactorysoftware/dreamfactory) project.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.1...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.2...HEAD
+[0.6.2]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.1...0.6.2
 [0.6.1]: https://github.com/dreamfactorysoftware/df-script/compare/0.6.0...0.6.1
 [0.6.0]: https://github.com/dreamfactorysoftware/df-script/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/dreamfactorysoftware/df-script/compare/0.4.0...0.5.0